0.5.06 - Added FLV viewing with Shadowbox - beta 2, Search for Extended Navigation
git-svn-id: https://plugins.svn.wordpress.org/photosmash-galleries/trunk@212359 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ ajax.php
@@ -546,22 +546,31 @@ class BWBPS_AJAX{
 						//Delete images that may be hanging out in the Meta
 						if((int)$row['wp_attach_id']){
 							$meta = get_post_meta((int)$row['wp_attach_id'], '_wp_attachment_metadata', true);
-																			
-							$folders = str_replace(basename($meta['file']), '', $meta['file']);
 							
-							if( is_file($uploads['basedir'] . '/' . $meta['file']) ){
-								unlink($uploads['basedir'] . '/' . $meta['file']);
+							
+							if(isset($meta['file']) && !empty($meta['file'])){
+																	
+								$folders = str_replace(basename($meta['file']), '', $meta['file']);
+								
+								if( is_file($uploads['basedir'] . '/' . $meta['file']) ){
+									unlink($uploads['basedir'] . '/' . $meta['file']);
+								}
+								
+								if(is_array($meta['sizes']['thumbnail'])){
+									$url = $uploads['basedir'] . '/' . $folders. $meta['sizes']['thumbnail']['file'];
+									if( is_file($url) ){
+										unlink($url);
+									}		
+								}
+								
+								
+								if(is_array($meta['sizes']['medium'])){
+									$url = $uploads['basedir'] . '/' . $folders. $meta['sizes']['medium']['file'];
+									if( is_file($url) ){
+										unlink($url);
+									}	
+								}		
 							}
-							
-							$url = $uploads['basedir'] . '/' . $folders. $meta['sizes']['thumbnail']['file'];
-							if( is_file($url) ){
-								unlink($url);
-							}		
-							
-							$url = $uploads['basedir'] . '/' . $folders. $meta['sizes']['medium']['file'];
-							if( is_file($url) ){
-								unlink($url);
-							}			
 							
 						}
 					}
